repositioning osc panel components; added params for env and added addional envs;

+++ gui/EnvelopeCurve.cpp
@@ -18,24 +18,24 @@ EnvelopeCurve::~EnvelopeCurve()
 
 void EnvelopeCurve::setSamples()
 {
-    
+
     sustainLevel_ = (96.f + sustain_) / 96.f;
-    
-    float samplesSection = getWidth()/4;
-    
+
+    float samplesSection = getWidth() / 4.0f;
+
     attackSamples = (attack_ * samplesSection/5) > 0.5f ?
     static_cast<int>(ceil(attack_ * samplesSection/5)) :
     1;
-    
+
     decaySamples = static_cast<int>(decay_ * samplesSection/5);
-    
+
     releaseSamples = (release_ * (samplesSection)/5 ) > 0.5f ?
     static_cast<int>(ceil(release_ * samplesSection/5)) :
     1;
-    
+
     // NOTE: the 2 at the end are responsible for the ending of the curve
     sustainSamples = getWidth() - (attackSamples + decaySamples + releaseSamples + 2);
-    
+
     samplesCounter_ = 0;
 
 }
@@ -78,7 +78,7 @@ void EnvelopeCurve::setReleaseShape(float releaseShape)
 float EnvelopeCurve::getEnvCoef()
 {
     float envCoeff;
-    
+
     // attack phase sets envCoeff from 0.0f to 1.0f
     if (samplesCounter_ <= attackSamples)
     {
@@ -135,7 +135,7 @@ float EnvelopeCurve::getEnvCoef()
     {
         return 0.f;
     }
-    
+
     return envCoeff;
 
 }
@@ -160,17 +160,17 @@ void EnvelopeCurve::paint (Graphics& g)
     backgroundFill.setOpacity(1.0f);
     g.setFillType(backgroundFill);
     g.fillAll();
-    
+
     Path curvePath;
     const int width = getWidth();
     setSamples();
-    curvePath.startNewSubPath(0, getHeight());
-    
+    curvePath.startNewSubPath(0.0f, static_cast<float>(getHeight()));
+
     for (float i = 1.0f; i < width; ++i) {
         curvePath.lineTo(i, (getHeight())*(1.013f - getEnvCoef()));
     }
-    
-    
+
+
     g.setColour(Colour(216, 202, 155));
     g.strokePath(curvePath, PathStrokeType(2.5f));
 }
